+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2] - 2022-02-06
+### Added
+- Custom link to for smoother routing
+- Font file preloading
+### Changed
+- NextSEO configurations for article tags and social media preview
+- Replaced Placeholder text on home page
+
 ## [0.4.1] - 2022-02-06
 ### Added
 - Added SCSS stylesheet to handle imported fonts